requirements.txt
itsdangerous==1.1.0
pytest==6.2.5
Flask_WTF==0.14.3
Flask_Bcrypt==0.7.1
WTForms==2.2.1
Flask==1.0.2
Flask_SQLAlchemy==2.4.1
Flask_Login==0.5.0
Pillow==8.3.2
secrets==1.0.2
